Move SectionToggle to components
