+++ front/app/containers/Admin/pagesAndMenu/components/SectionToggle/index.tsx
